pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-sentinel"
version = "0.1.0"
description = "SRE Incident Triage Environment for OpenEnv"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
inference = [
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "httpx>=0.25.0",
]

[project.scripts]
server = "server.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["openenv_sentinel*", "server*", "scenarios*", "tools*", "grading*"]
